+++ packages/orq-rc/src/models/operations/createcontact.ts
@@ -134,7 +134,7 @@ export const CreateContactResponseBody$inboundSchema: z.ZodType<
   created: z.string().datetime({ offset: true }).transform(v => new Date(v))
     .optional(),
   updated: z.string().datetime({ offset: true }).default(
-    "2026-02-18T12:44:37.542Z",
+    "2026-02-18T16:59:05.087Z",
   ).transform(v => new Date(v)),
 }).transform((v) => {
   return remap$(v, {